Usage

ProjectAssigner has a single function, match for matching students to projects. In Julia, typical usage is as simple as:

import ProjectAssigner
ProjectAssigner.match(students="students.csv", projects="projects.csv", output="output.csv")

Similarly, in Python, typical usage looks like

from julia import ProjectAssigner
ProjectAssigner.match(students="students.csv", projects="projects.csv", output="output.csv")

Note that all arguments are keyword arguments. See the docstring below for information on the function options and the Data Format section for details about the input.

ProjectAssigner.match — Function

ProjectAssigner.match(;students, projects, [output], [force], [optimizer])

Match students to projects and return a dataframe with the assignments.

All arguments are keyword arguments, and the function is not exported, so it must be called with the module explicitly ProjectAssigner.match(...).

Example

julia> ProjectAssigner.match(students="students.csv", projects="projects.csv")

Keyword Arguments

students: A DataFrame or csv filename String specifying information about the students. See the Data Format section of the documentation for more information.

projects: A DataFrame or csv filename String specifying information about the projects. See the Data Format section of the documentation for more information.

output::String: optional name for a csv file that the assignments will be written to.

force: optional vector of student-project pairs to force matches for, e.g. force=["Bob"=>"Airplane"] will force student Bob to be assigned to the Airplane project.

optimizer::Function: optimizer factory to use for the optimization, e.g. optimizer=Gurobi.Optimizer will use Gurobi instead of the default GLPK

Scaling Issues

Due to the exponential costs in the objective, the optimization problem may be poorly-scaled for large classes. This can result in the default GLPK optimizer producing poor results even though it thinks that it has solved the problem. More powerful commercial solvers such as Gurobi can handle this much better.